\name{load_rosters}
\alias{load_rosters}
\title{Load Team Rosters}
\usage{
load_rosters("")
}
\description{
Input options: "mens" "womens" or "mixed" - loads roster data from {https://github.com/jbrooksdata/2022-olympic-curling}
}
\examples{
load_rosters("mens")
load_rosters("womens")
load_rosters("mixed")
}
