\name{load_standings}
\alias{load_standings}
\title{Load Standings}
\usage{
load_standings("")
}
\description{
Input options: "mens" "womens" or "mixed" - loads team standings and stats from {https://github.com/jbrooksdata/2022-olympic-curling}
}
\examples{
load_standings("mens")
load_standings("womens")
load_standings("mixed")
}
