\name{load_player_games}
\alias{load_player_games}
\title{Load Player Game Logs}
\usage{
load_player_games("")
}
\description{
Input options: "mens" "womens" or "mixed" - loads player game logs from {https://github.com/jbrooksdata/2022-olympic-curling}
}
\examples{
load_player_games("mens")
load_player_games("womens")
load_player_games("mixed")
}
