\name{load_team_games}
\alias{load_team_games}
\title{Load Team Game Logs}
\usage{
load_team_games("")
}
\description{
Loads men's, women's, or mixed doubles game logs from {https://github.com/jbrooksdata/2022-olympic-curling}
}
\examples{
load_team_games("mens")
load_team_games("womens")
load_team_games("mixed")
}
